requirements.txt
yfinance
pandas
numpy
matplotlib
MetaTrader5
python-decouple
websocket-client